Initial commit for ArkoseRenderer 2.0!
The previous repository had a very messy history and with a lot of heavy binary assets in it, so I decided to create a clean repo. The old repo is available at https://github.com/Shimmen/ArkoseRendererThesis.

Additionally a large refactoring has been made between this and the old repo. Mainly, a better build process, a new resource system, and better directory structure.