Add a first iteration of bloom
The bloom is currently blurred with some bias towards
the edges of the screen, which is a bug. This should be
fixed. Also, the barriers controlling the R/W for the
iterative upsample & downsample passes needs to be
more precise so we don't stall the whole GPU just for
that.