Use binding set desc. set layouts for GPU states
I.e. instead of deriving descriptor set layouts from the shader data,
we use the layouts as described by the binding sets. This makes it
so that we can make a set work for all possible different shader stages
while before it was very tricky.

The one problematic thing about it is that we now require the
binding sets to be listed in the right order already at gfx/compute/rt
state creation. This is alright really, but it needs tightening up.