Ray tracing refactor (support for KHR extension)
We now have support for both RTX (Nvidia) and KHR (Khronos)
ray tracing extensions, and in theory we should be able to swap
between the two with no issues.

There are currently some issues with device lost for KHR ray tracing,
but there are no validation error and no obvious issue so I will check
it in as is and work on those issues later. We also need to rename the
old code path and make it more explicitly an extension and not a core
part of the vulkan backend.