Add simple TaskGraph and ParallelFor
The ParallelFor is implemented in terms of the TaskGraph. These two
features should help us speed up the engine significantly!

The TaskGraph does not currently suport task dependencies, but as
it's not something I see an immediate need for I've just delayed that
implementation and it will definitely be added later.