Massive scene refactor
Split GPU specific Scene data into a GpuScene class. The idea is that
Scene is a generic type for a scene in its whole, while classes like
GpuScene and PhysicsScene handle specific cases. While there is no
clear purpose of this right now, I see two main benefits, both of which
will be done soon:

 1. It will make it so much easier to separate the idea of a GPU render
    object and an uploaded texture, compared to just some description
    on the CPU of what it should be (see e.g. environment map, for
    which I already did something like this). In time this will allow us
    to remove the `rebuildGpuSceneData` function and have a more
    dynamic way of handling all that.

 2. It will be easy to create a physics abstraction for the objects we
     have in a scene, whithout mixing it up with rendering.

There are now quite a lot of bad access pattern (see scene.scene(). ...
where the first scene is a GpuScene and the second one is Scene), but
these can be cleaned up over time.