Add basic Tracy Vulkan integration
It currently only captures the overall frame and the precence of the
named Nodes, but not their names. Ideally we'd like to hook into the
begin/end debug label functionality used by the command list, but
due to how Tracy requires 1) constant stings, and 2) a single scoped
marker (instead of a separate start and end) we can't really do that
right now.