Refactor TracyVk integration to use single context
This makes the timeline in Tracy nice and readable, and it makes more
sense to have a single context per queue. The documentation doesn't
explicitly say that the command list you pass in must be the same or
not be the same, but turns out that it doesn't need to be, making this
easier.