Add support for async loading material textures
Still a lot of work to make it great, but we can at least now load
big scenes, like the Intel Sponza very quickly and without blocking
on image load.

Our task graph is also not great yet and is mostly FIFO with some
randomness, so it's easy to ovecommit too much work and then other
work not being scheduled until all previous is done. This happens with
Texture::createFromImagePathSequence right now, so to avoid a large
stall waiting for image path sequence textures to load in we keep it
single threaded for now. Later, we need 1) priorities, and 2) a faster
way of loading array textures, e.g. a single binary blob.