Rename src/ to arkose/
We will soon start adding tooling etc. and it makes sense to keep that
code separate, but with a name as src it implies all source can be found
in that directory. This naming makes it more clear that it's where the
arkose code, i.e. the renderer, lives. It would be nice to also move the
shader/ directory to in there but that would require some more pipeline
work.