Begin work on Depth of Field effect
For now we just calculate the CoC, and we're not yet blurring anything.
We also just have a slider for selecting focus depth. Later we probably
also want to be able to click on a point in the world and figure out
the depth for it to use as focus depth.