Remove USE_DEBUG_TARGET option from the DDGI node
Now when practically all RT shaders are shared between all the
nodes we can use the RTVisualisation node and be fairly sure
that the DDGI node sees the same thing and the debug target
has no purpose any more. It was also quite messy and didn't work
so well, so getting rid of it is very good too.