Add flatbuffers-based asset (& serialisation) system
It's still work in progress with a lot left to do. This should at least
let you read and write images, materials, and static meshes to disk
in a binary (flatbuffers) format. We also want to support json later
which is doable through flatbuffers as well.

Right now we have a separate GltfLoader which loads into these new
formats. This works but it can't yet send these loaded assets to the GPU
for rendering, since the GpuScene is only aware of the old types. We
need to create runtime versions of the assets which could e.g. keep a
back-reference to the asset, but mainly be very light-weight and only
deal with GPU stuff.