Add example level in the LevelAsset format
Note that the mesh asset paths are manually set for now. Later we will
have to implement some quick-and-dirty import from e.g. glTF to our
asset types on first run. E.g., if the path refers to a glTF, see if
there exist an imported variant of it and use that or make it.