Add masked material support for prepass
This is needed for directional shadow projection, as it needs to know
all opaque pixels on screen for projecting correctly.

We might want to avoid creating too many shaders if we can instead
use permutations of e.g. the forward shaders, but we also don't want
to clutter the shader code with permutation logic. There is a balancing
act, and for now I think this is okay since the prepass code is quite
short and simple to duplicate.

We also want to consider if the large amount of vertex layouts we're
creating to support all the different shaders is reasonable. We probably
want to combine all of them into two streams: a position only + the rest