Handle relative includes also for finding last edit timestamp
We previously added support in the Shaderc include handler for it,
meaning we did compile it correctly before. However, the shader hot
reloading or the shader spirv cache would not know that a shader would
need compiling if a file included with a relative path would be changed.
This is fixed with this change.

Also improved the error reporting in case one or more files were missing
(i.e. can't be found on disk) in the include tree.