Make GI composition almost correct
We still don't use the specular BRDF for reflections as we don't yet
have the PDF at hand. This will need to be done at some later point.

Diffuse GI is now sampled in the composition shader instead of in
the forward shader. This makes the correct-ish indirect lighting easier
to do.