Stream in textures w.r.t. memory budget per frame
Instead of just uploading up to X textures per frame we instead upload
as many as we can until we fill up our budget. For now we always stream
in all mips of an image (it's just simpler that way) but it's not so
hard to build a proper mip-based texture streamer from what we have now.