Remove all skinning data & logic from meshlet code
The idea is that the meshlets only will care about post-skinned i.e.,
from it's point of view, stratic meshes, so it will never have to do any
skinning itself. (Notable exception is velocity rendering, which will
need to at least be aware of where the vertices were last frame, with
means we need to keep track of last frame's skinned result as well).