Upgrade to Tracy v0.9.1
We're now using the provided CMakeLists.txt and we're now also able to
build the tracy client as fully a separate library, which wasn't
previously possible, as far as I know. However, there seems to be some
issue with TRACY_ON_DEMAND so we will have to do without it for now.