Restructure into arkcore lib and arkose executable
This change moves everything that isn't specific to the runtime code to
the new arkcore / ArkoseCore static library target, which ArkoseRenderer
the executable depends on. This will allow reuse of assets and more for
other targets, for example, various tools and asset processing.

For now, the arkcore files can be included as-is, meaning that all
exisiting #include for these files can be left unchanged. Later we might
want to put it behind the "arkcore" include path so that it's always
clear where the file is located, and avoid the horrible path ambiguity
that unreal engine has. On the other hand, for this size project maybe
this isn't so bad.. That's a decision for a later time. For now it makes
the transition nice and easy.

To accomodate this refactor some functional changes had to made:
 - LevelAsset no longer refers to Camera and Light types directly, as
 they contain a lot of runtime data. Instead, we now have CameraAsset
 and LightAsset which then the Camera and Light classes can be created
 from. This is the largest actual code change from this refactor.
 - Arkose can now compile the different backends conditionally, so we
 can only compile the Vulkan backend for macOS etc.
 - I can't get the ARKOSE_LOG macros to work properly with clang, so for
 now they are no-op until I can figure out how to compile them..