arklib: move to deps/ directory
This will make the distinction beween arklib and arkcore more clear,
as arklib completely generic and not arkose specific (but yeah it's got
the name...) while arkcore is a core library for arkose specifically.
The idea is to move arklib to its own git repository at one point in the
future, so this move makes sense in the end.