Fix ARKOSE_LOG macros for clang (& gcc) compilation
It was missing `__VA_OPT__(,)` which is required for when we have no
arguments, but for whatever reason MSVC didn't care about that. It's
possible that MSVC is happy with __VA_OPT__(..) but I'm not entirely
sure yet, so will have to test before I make it compile for that too.