Add global mip bias and apply to forward passes
Note that we still have to do something similar for the visibility
buffer case where we manually specify gradient, which ironically
makes it quite tricky to apply a simple bias. If not for e.g. anisotropy
we'd be able to calculate the lod value and bias it, but with all those
details in mind it's actually pretty tricky to make something robust.
Will have to think about that one.