Make the subsurface scattering pass mostly correct
It's incorrectly using scene color in place of irradiance, which will
cause overblurring of specular highlights. See the comment where
we combine sample in the shader.

The implementation also has no support for any kind of SSS-masking,
so hair etc. will be equally blurred if defined in the same material,
which doesn't look that good. However, "good" assets should probably
ensure they are separate materials anyway.