Move bloom to after DoF
Not sure if either is more correct, but I think having bloom as late as
possible in general makes sense, so the pixels of the "final" image is
allowed to contribute to the bloom. In this case, if DoF blurs out some
bright spot in the background we will now effectively bloom it with the
the size we're seeing on screen, not what it actually is. Is this energy
preserving? Maybe not :^)