Rename "GI compose" to "Lighting compose"
For now it does the same but I plan on having it do more heavy lighting
composition work in the future and let the forward lighting passes write
out the different results in separate render targets. This will ensure
we can e.g. do SSS on the diffuse irradiance only and then combine
everything together in the end.