Use diffuse irradiance for subsurface scattering
We're currently only including direct diffuse irradiance for subsurface
scattering, which is not 100% correct, but I think in most cases the
indirect shouldn't contribute that much irradiance compared to the
direct, so in practive it might not change much.

Also, I noticed a very slight difference in intensity of the overall
lighting when doing separate diffuse lighting from before - I hope this
is just some precision issue and not some correctness thing, but either
way it's not ideal.. I hope I can resolve this in the future, or at
least find the root cause, because now I'm not quite sure why.