Reset sharpness to default for SSSDemo
Now when the subsurface scattering (correctly) only blurs the diffuse
irradiance and not the whole scene color we get nicely sharp highlights
still, and the stubble etc. is not completely blurry, so we don't need
such intense sharpening for it to look nice. Also, I had the roughness
of the head/skin material way too high before, which contributed to
very little highlights; that too is now much better.