Make Input class system/platform independent
Note that the key, button, action, modifiers values are identical to GLFW's
but that's purely for the sake of keeping it nice and standard. There's not
actually a strict dependency, just something to keep in mind for when
adding more systems.

Slightly broke gamepad support as part of this refactor, as it was tied to
GLFW and was very hacky anyway (and never used by me). Now we're in
a good state to add it back again, properly this time.