Convert ComputeState to use StateBindings & remove `bindSet` function
There is one important exception, where we do actually still call bindSet
manually (albeit indirectly) which is `bindTextureSet` in DebugDrawNode.
See the comment in CommandList.h for the motivation of this.

The reason we're doing this in the first place is to make it simpler and
more compatible with other graphics APIs, such as D3D12.