D3D12Backend: handle window resize
It's quite buggy and doesn't play very well at all with TracyD3D12, but it
technically sort of works. It appears as if you very quickly resize it sometimes
crash but otherwise it is okay. Possibly we're not correctly waiting for an idle
frame and recording things on the one we just destroyed resource for.. needs
more investigation, but I'll check this in for now as it technically works.