D3D12: Fix swapchain resizing behavior
It no longer crashes 50% of the time we resize :^)
This was fixed by actually explicitly track which the next
swapchain buffer index is, so we can start rendering for
that one for the following frame.

To get that functionality we needed to "upgrade" to the IDXGISwapChain4 API, and some other changes related
to that.