Add cmake script for DirectXAgilitySDK nuget package
It will grab the nuget package of the specified version and setup
an interface library for the headers (there's technically one source
file as part of the nuget package but we don't need it for now).

The DLL files are not part of the build process at all, as they are
fully loaded at runtime by the Agility SDK runtime of a modern
version of Windows 10+, so we just export their path so they
can be copied to where they are needed by the caller.