Remove buffer memory hints
Implementing the D3D12 backend I came to realize that the memory hints
are both not very useful and in some cases also ignored and/or ambigous
in terms of what they do. In this change I remove the buffer memory
hints, but I add a few more buffer usages (maybe should be renamed to
type) and it's overall simpler to reason about and better in terms of
how the hardware is implemented.

As part of this change it was also convenient to make use of the
D3D12MemoryAllocator for buffers, so that change is included.

There's a slight VRAM memory leak from the D3D12MemoryAllocator in this
change, just FYI.