Add optional `stride` to Buffer
This can be used by some implementations that need an explicit stride
when binding buffers or creating buffer views. For example, D3D12 needs
it to create a view for a regular- or structured buffer. To make this
enforcable across all backends we will require a buffer to have stride
specified if it is to be used as a storage buffer in a shader binding.
This is not yet enforced but will be in the next commit.