D3D12: Define static samplers for each combined image sampler in a binding set
This won't work if we have a bindless array of thousands of such combined
image samplers, but for now this will work. I think we might just go with separate
sampler objects for bindles, and enforce that somehow. I still want to keep the
combined ones though, for their convenience.