Path tracer: rename ray tracing shaders
For now we only have a single hit group but later we'd likely want
multiple of them for different types of materials, and then we'd want
to reuse these shaders for most types. In the short term we might also
want to make the same shaders & hitgroup handle all materials with
dynamic branching (it's not optimal but it's easy to implement as a
first step).