Add new API for constructing a shader binding table more explicitly
We can now specify hit groups and miss shaders explicitly with an index
which matches with how the instances are created by the GpuScene.

Add `valid()` checks for ShaderFile and ray tracing HitGroup. Also make
it possible to make invalid hit groups with the default constructor, as
it makes it possible to dynamically grow the vector of hit groups.