Use multiple hit groups for ray tracing
The GpuScene will now put instances at shader binding table offsets
that are relevant for what's needed and existing shader binding tables
are modified so there's definitely a shader there (or else we're in
undefined behaviour territory). Later we probably also need to handle
multiple different BxDFs with these offsets but for now it's enough to
separate opaque/masked/translucent, as we only have a single (usable)
BRDF for each.