Path tracer: add support for polished glass materials
It currently has a few flaws. First, it always assumed we pass from air
to glass or the other way, but never between two non-air interfaces. It
also always assumes that the camera is not inside glass. Finally, we
only light surfaces that are front facing, so inside of glass objects
will not recieve any direct lighting.