Change shader compilation to on-demand instead of upfront
i.e. we compile or load the cache for a specific type (SPIR-V or DXIL) when requested,
instead of doing it all upfront when a shader file is first added. Now the `ShaderFile` constructor just "registers" the file but nothing is compiled or loaded at that time.

This also means we don't try to transpile GLSL/SPIR-V to HLSL for files we know can't
transpile yet anyway.