Collect named constants in a shader metadata file stored in the shader cache
For now it's not directly used, besides for verifying that named constants are valid,
non-overlapping beteen all shader files of a shader. Later this will also be used to
look up named constants in runtime when setting them. We currently already do
something similar, but we 1) construct them lazily, which I don't love, and 2) do it
in a way that it's hard for e.g. the D3D12 backend / DXC-compiled shaders to use
that data. This new way solves both of these issues.