Improve glTF asset image handling
Previously we had assumed there was a 1:1 texture to image mapping,
but that's not actually the case. In the future we may want to separate
sampler info etc. (i.e. what's in the glTF "texture") from the image
data (i.e. what's in the glTF "image") but for now it's all the same,
so we need to actually generate an arkimg for each "texture".

Also some related fixes & improvements:
 - remove the userData field as it was unused
 - copy the asset name in `createCopyWithReplacedFormat`
 - handle a few error cases in `generateMipmaps`
 - give reasonable names to images stored within a glb et al.

All in all these changes allow us to load in more assets successfully.