Add black bar post-process rendering
It won't change the depth values to improve culling etc. but I think that's alright. If you're
expected to be rendering with these bars at all times then I think it definitely should, but
in this case it's just a matter of visual flair for a limited amount of time.