Add support for color grading
The color grading is controlled with a 3D LUT, which can be loaded from a .cube file.
For now, by default we load the "identity" LUT which means it's effectively a no-op
color grade, but there's roughly a 0.1ms cost to applying this LUT. Later we should
probably consider converting the LUT format to RGBA16F or something lighter, and
also making better UI to indicate what kind of color grade we're doing etc.