Remove `ARKOSE_LOG_FATAL` macro and replace with `ARKOSE_LOG(Fatal, ...)`
It appears that MSVC now understands that `ARKOSE_LOG(Fatal, ...)` has noreturn behaviour,
so it understands we won't progress past it e.g. with a definite nullptr. Instead, with more
warnings enabled it complains at `ARKOSE_LOG_FATAL` as it has a call to `exit()` after a
previous noreturn, leading to warnings about unreachable code, hence this change.