Make TAA node work (only) before tone-mapping
This is more in line with what other engines do. Ideally, we'd even put
TAA before bloom etc. so that we avoid flickering/strobing in the bloom
results, but this will do for now. The reason I'm doing it now in the
first place is just to make things easier for future HDR output work,
where we want to be able to do TAA followed by HDR-output tone map.

Also clean up the TAANode class with more modern Arkose code
conventions.