Vulkan: always use persistently mapped buffers when applicable
That is, any buffer which we can/will map & unmap - as opposed to the device-only ones, which we'd copy to from a staging buffer - are now persistently mapped. This potentially
saves some performance, but nothing that I've been able to measure.

However, while capturing a frame for RenderDoc `vmaUnmapMemory` can take up to 0.5s
per call, which massively slows down the capture process. By moving to persistently mapped
buffers we can reduce the capture time from ~30s down to <1s, which is much nicer. This is
the true reason why I'm actually changing over to using persistently mapped buffers :)