Add padding-only vertex components
This is useful for when using the Arkose default defined vertex buffers, such as the
non-position vertex buffer - which contains texcoords, normals, and tangents - but only
need one or a few of the components. Unless we handle this case the vertex assembler
in the hardware will still load all the vertex data, which is redundant. The shader isn't
consuming these components, but I presume that information is not available to the
right hardware systems at the time needed, as Vulkan validation generates performance
warnings if we don't handle this correctly. Note that padded vertex components don't
count as real components, so the first non-padded component will have index 0 when
referenced from within the shaders.