Write bent normals out to an intermediate scene texture
This is a pretty bad way of doing this, as we're essentially just copying a bunch
of data around with no benefit to it. Long term we really have to commit to
one of the paths: forward or visibility buffer.. Speaking of, currently only the
visibility buffer mode actually writes out valid values.

We're also not handling default values / no bent normal texture very well.
We likely want to enforce that every texture has one, or simply fall back on
the standard normal map. However, this requires us to write out tangent
space bent normal maps, not object space like we do now. Much to do!