Vulkan: refactor imageless framebuffers
Instead of explicitly saying that a render target is imageless we can now derive it from the attached textures. If one of the textures is
the backend's special "placeholder swapchain texture" then the
render target will use an imageless framebuffer & we handle the
binding accordingly. This doesn't make a great deal of difference
for now, but will allow us to get rid of the "window render target"
and instead let anyone create a render target for the swapchain
in the normal method.

This change also cleans up the swapchain image layout transitions.