Add OutputNode which combines the Tonemap & Final nodes into one
This allow us to handle both sRGB & HDR output, as we need to treat the
color quite differently on the way to the display depending on the two.
The main downside is that we can't easily support FXAA, since to run it
as its own separate pass we need to output sRGB. However, in practice,
we don't really use FXAA any more any way, and there are ways to make
it work in the future if needed, so I'm happy to bread this for the time
being. The CAS node is also somewhat broken with this, as it also
expects to run as its own pass, but we can live without it for now, and
I do suspect we can make this too work in the future if we want to.