Ensure we never have null tangents in meshes
This seems like quite the hack, and I'm not sure why the MikkTSpace library returns such
a tangent vector. Possibly to indicate it's a degenerate triangle or something akin to that?
However, we need "valid" tangents (length == 1) for all triangles so let's at least write
something valid in these cases.