Vulkan: remove support for the 'VK_NV_ray_tracing' extension
We now only support the KHR-family of ray tracing extensions, and it looks like Nvidia
themselves are also on board this standardization. I'm pretty confident to say we're not
loosing out on any future stuff from Nvidia by not supporting it, as it would likely just be
just another extension depending on the KHR ones instead of the NV one.