Add render target attachment clear API
Note that this is separate from texture clears as it can potentially be done more
optimally as an attachment, with various HW render target color compression stuff.

This was intended to be used for the local light shadow draw tiles but ended up
not doing that in the end, so just adding it here for future use.