Remove broken instance-level culling from meshletTaskSetup
It was intentionally broken when adding support for meshlet culling
and rendering of shadows with meshlets. Soon enough I will add proper
multi-view culling of objects (plus LOD selection and all that fun) so we'll
get this back, better than ever, in time. For now though, let's rip it out.