Re-add the DDS-as-image-source path
It's a bit of a misnomer, as DDS is now our native image format, and it requires some code
duplicaton as-is. However, removing this path breaks the import of the Bistro asset, which
uses DDS asset files. So let's solve this properly at some future time, but for now, it works.