Always normalize quaternions!
These changes fix various related issues with bad quaternions, like light directions
not behaving like we'd expect. With this change we pretty much never trust a quat
at face value and will always normalize when serializing or assigning to a Transform,
or in any place where we'd expect normalized quaternions.