arklib: implement a simpler version of `quatToMatrix`
This simpler version is probably more expensive than the old one (which I had copied
from glm's implementation) but it also guarantees that the matrix axes are all length
one, which the old one did not.

Thinking about it though, this may actually be just a side-effect of the lack of quat
normalization which was prominent in the Arkose codebase, which is now fixed..
Either way, this new implementation appears to be more resilient against this.