Add explicit `Sampler` resource type
Previously we always just defined samplers as part of the Texture, however, this
doesn't scale well when going bindless in D3D12 as there's a limit to the number
of samplers making it hard to use combined image/samplers like we do in Vulkan.
The idea is to keep using combined image/samplers for most cases, but pass along
explicit samplers for bindless textures, and bind those textures as regular texture
descriptors.

There's a fair amount of duplication in how we create the samplers for a texture
and for these explicit samplers, but hopefully we can reduce this going forward.
Possibly let the textures defer the sampler creation to the sampler types.