Add a clearcoat to the default BRDF material model
This break the subsurface scattering (skin) implementation, as we're now no longer
writing out a separate diffuse irradiance value from the visbuf shading. This will need
to be fixed up by making a separate skin BRDF which does that. But I'll break it for the
time being, so we get clearcoat working.

Clearcoat is not yet handled correctly in the path tracer shaders. We handle it for the
direct lighting but not for the material sampling & evaluation. To do!