Re-implement support for Skin BRDF through the v-buffer path
We leep specular & diffuse separate, as we need separate diffuse irradiance to
simulate subsurface scattering for the skin in a later stage. However, now the
target is only used for skin, no other diffuse lighting is put there. Thus, we also
rename the options a bit for the lighting compose pass.

Just like before, we still don't support the Skin BRDF through the regular forward
pass, as there are some v-buffer specifics in the SSSS shader.