Improve AssetCache API so it's impossible to double-create the same asset
Example flow that was previously possible but is no longer:
 1. get(x) -> nullptr
 2. make x1
 3. get(x) -> nullptr
 4. make x2
 5. put(x1)
 6. put(x2) -> x1 now gets deleted!

By making `getOrCreate` the only way to get an entry from the cache, we now avoid this situation entirely.

We still allow put:ing arbitrary data in the cache, but that should be safe as we do have a mutex lock for
any cache access.