Move directional light shadow mask ownership to the GpuScene
This is step 1-ish of making a bit more of a unified system for shadow maps, ray traced shadows,
and all the logic to manage this.

Also change terminology: "projected shadow map" -> "shadow mask", for a render resolution sized shadow texture which can be sampled directly with a screen coordinate.