Add new concept of `ExternalFeature` and reimplement DLSS integration with it
ExternalFeatures can be created & evaluated in a generic way. There's a few void* here and there,
but it's not too bad. All in all I'd say the DLSS usage may be a little less clean, but we can get rid
of the silly UpscalingState, and mainly, we can now implement other external features without
adding more besboke APIs.