Change how DLSS support & render resolution are queried for
In a way we make it less generic, by exposing DLSS-specific functionality to the Backend, but I think
it's a reasonable way to be able to query for support etc. for it. At the end of the day, what we care
about is whether the backend supports DLSS or not, so asking for it directly is fine.