Fix GPU crash caused by copying vertex data out-of-bounds
Only affects very large scenes, for obvious reasons if you look at the changes,
but it should now be enough for even very large scenes, and with better
validation.

There's also something weird going on with the UploadBuffer where if it's not
big enough we get crashes.. not sure why, we got plenty of validation to ensure
we're not OOB with that.. More to discover.