Tools: add a new ShaderCompilerTool using slang
Just a work-in-progress at the moment, the idea is that we pass in some
kind of shader spec which specifies a set of shader files, their types,
and a set of defines shared by all. For now it's just a little test
case of using slang to compile glsl code.